split decrypt tests to separate file
